moved the file-tree module from the navigator to the filesystem extension
