+++ src/filesystem/browser/file-tree/file-tree-container.ts
