+++ src/filesystem/browser/file-tree/file-tree-model.ts
