+++ src/filesystem/browser/file-tree/file-tree-widget.ts
