+++ src/filesystem/browser/file-tree/file-tree.ts
