+++ src/filesystem/browser/file-tree/location-renderer.ts
