+++ src/filesystem/browser/file-tree/location-service.ts
